requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
autopep8==1.5.4
Django==3.1.2
django-environ==0.4.5
django-extensions==3.0.9
django-url-checks==0.2.0
djangorestframework==3.12.1
flake8==3.5.0
flake8-bugbear==18.2.0
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8-naming==0.7.0
pycodestyle==2.4.0
pydocstyle==5.1.1
pyflakes==1.6.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
